Various QoL improvements over the original
When an order fails (error returned by Binance API) the log now correctly reflects this
Order quantity now always rounds up to nearest denomination of a coin allowed on Binance. This means DCA orders of the minimum amount (10 UDST) will no longer fail for falling below the minimum trade amount where the quantity has been rounded down
Bot now uses spot market orders instead of margin orders as in the original
Test trades now use the Binance API test trade request, giving authentic error codes for testing purposes
Instead of using a sleep command for the full duration of the DCA wait period, I've incorporated a check every 15 minutes to see if enough time has passed since last trade. This way the script is crash and reboot resistant. In the original a restart of the script would mean an instant buy of all coins. This method allows you to restart and it will still wait the full duration. Also logs each attempt to clearly show script is still running. If DCA period is less than 15 minutes, the sleep time is reduced to the DCA period
Buys are only logged if order was successfully placed on Binance, ensuring records are accurate. If a buy order failed in the original it would still be logged in the history